changes alla pages class name
