--- UI/Pages/CheckTransactions.cs
@@ -1,32 +1,0 @@
-﻿using bank_app.Managers;
-using bank_app.Models.Users;
-using bank_app.Utility;
-using bank_app.Utility.Components;
-using bank_app.Utility.UI;
-using bank_app.Utility.UI.Components;
-using bank_app.Utility.UI.Invokables;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace bank_app.UI.Pages
-{
-    internal class CheckTransactions : Page
-    {
-        internal override UIComponent LoadPage()
-        {
-            Console.CursorVisible = false;
-            var navbar = new Navbar(new List<NavbarItem> { new NavbarItem("Home", PageType.AdminDashboard), new NavbarItem("Create User", PageType.CreateUser),
-                new NavbarItem("Transactions", PageType.CheckTransactions), new NavbarItem("Rates", PageType.UpdateRates)});
-            // Create a new Invokable object with the method that is going to run once the 'Login' button is pressed
-            // Create new 3x3 grid
-            Grid grid = new Grid(3, 3);
-            grid.AddGridComponent(0, 1, navbar);
-            grid.AddGridComponent(1, 1, new Text("Home"));
-            // Add grid to layout and set rounded border style
-            return new Panel(grid, LayoutBorder.Heavy);
-        }
-    }
-}
